+++ lib/ood_core/batch_connect/templates/vnc_container.rb
@@ -1,4 +1,5 @@
 require "ood_core/refinements/hash_extensions"
+require "securerandom"
 
 module OodCore
   module BatchConnect
@@ -7,15 +8,20 @@ module OodCore
 
       # Build the VNC template from a configuration
       # @param config [#to_h] the configuration for the batch connect template
-      def self.build_vnc(config)
+      def self.build_vnc_container(config)
         context = config.to_h.symbolize_keys.reject { |k, _| k == :template }
-        Templates::VNC.new(context)
+        
+        unless context.key?(:container_path)
+          raise JobAdapterError, "You are missing the configuration 'container_path' for a vnc_container template."
+        end
+
+        Templates::VNC_Container.new(context)
       end
     end
 
     module Templates
       # A batch connect template that starts up a VNC server within a batch job
-      class VNC < Template
+      class VNC_Container < Template
         # @param context [#to_h] the context used to render the template
         # @option context [#to_sym, Array<#to_sym>] :conn_params ([]) A list of
         #   connection parameters added to the connection file (`:host`,
@@ -53,21 +59,48 @@ module OodCore
         # @option context [#to_s] :vnc_clean ("...") script used to clean up
         #   any active vnc sessions (assumes you don't modify `:before_script`
         #   or `:clean_script`)
+        # @option context [#to_s] :container_path ("vnc_container.sif") the path
+        #   to the container with VNC
+        # @option context [#to_s] :container_bindpath ("") paths to bind into
+        #   the container with VNC
+        # @option context [#to_s] :container_module ("singularity") the module
+        #   that loads Singularity or Apptainer with Lmod. Supports versions (i.e.
+        #   apptainer/1.10). If Singularity or Apptainer are installed at a 
+        #   system level (i.e., no module loaded to activate), set this to an
+        #   empty string.
+        # @option context [#to_s] :container_command ("singularity") the 
+        #   singularity or apptainer execution command
+        # @param instance_name (uuid) a name for the instance
         # @see Template
+
         def initialize(context = {})
+          @instance_name = SecureRandom.uuid
           super
         end
 
         private
           # We need to know the VNC and websockify connection information
           def conn_params
-            (super + [:display, :websocket, :spassword]).uniq
+            (super + [:display, :websocket, :spassword, :instance_name]).uniq
           end
 
           # Before running the main script, start up a VNC server and record
           # the connection information
           def before_script
+            container_path = context.fetch(:container_path, "vnc_container.sif").to_s
+            container_bindpath = context.fetch(:container_bindpath, "").to_s
+
             <<-EOT.gsub(/^ {14}/, "")
+
+              # Load #{container_module}
+              echo "Loading #{container_module}..."
+              module load #{container_module}
+              export #{container_command.upcase}_BINDPATH="#{container_bindpath}"
+              export INSTANCE_NAME="#{@instance_name}"
+              export instance_name="#{@instance_name}"
+              echo "Starting instance..."
+              #{container_command} instance start #{container_path} #{@instance_name}
+              
               # Setup one-time use passwords and initialize the VNC password
               function change_passwd () {
                 echo "Setting VNC password..."
@@ -75,11 +108,12 @@ module OodCore
                 spassword=${spassword:-$(create_passwd "#{password_size}")}
                 (
                   umask 077
-                  echo -ne "${password}\\n${spassword}" | vncpasswd -f > "#{vnc_passwd}"
+                  echo -ne "${password}\\n${spassword}" | #{container_command} exec instance://#{@instance_name} vncpasswd -f > "#{vnc_passwd}"
                 )
               }
               change_passwd
 
+              
               # Start up vnc server (if at first you don't succeed, try, try again)
               echo "Starting VNC server..."
               for i in $(seq 1 10); do
@@ -87,13 +121,14 @@ module OodCore
                 #{vnc_clean}
 
                 # for turbovnc 3.0 compatability.
-                if timeout 2 vncserver --help 2>&1 | grep 'nohttpd' >/dev/null 2>&1; then
+                if timeout 2 #{container_command} exec instance://#{@instance_name} vncserver --help 2>&1 | grep 'nohttpd' >/dev/null 2>&1; then
                   HTTPD_OPT='-nohttpd'
                 fi
 
                 # Attempt to start VNC server
-                VNC_OUT=$(vncserver -log "#{vnc_log}" -rfbauth "#{vnc_passwd}" $HTTPD_OPT -noxstartup #{vnc_args} 2>&1)
+                VNC_OUT=$(#{container_command} exec instance://#{@instance_name} vncserver -log "#{vnc_log}" -rfbauth "#{vnc_passwd}" $HTTPD_OPT -noxstartup #{vnc_args} 2>&1)
                 VNC_PID=$(pgrep -s 0 Xvnc) # the script above will daemonize the Xvnc process
+                echo "${VNC_PID}"
                 echo "${VNC_OUT}"
 
                 # Sometimes Xvnc hangs if it fails to find working disaply, we
@@ -135,9 +170,10 @@ module OodCore
               #{super}
 
               # Launch websockify websocket server
+              module load #{container_module}
               echo "Starting websocket server..."
               websocket=$(find_port)
-              #{websockify_cmd} -D ${websocket} localhost:${port}
+              #{container_command} exec instance://#{@instance_name} #{websockify_cmd} -D ${websocket} localhost:${port}
 
               # Set up background process that scans the log file for successful
               # connections by users, and change the password after every
@@ -156,9 +192,11 @@ module OodCore
           def clean_script
             <<-EOT.gsub(/^ {14}/, "")
               #{super}
+              module load #{container_module}
 
               #{vnc_clean}
               [[ -n ${display} ]] && vncserver -kill :${display}
+              #{container_command} instance stop #{@instance_name}
             EOT
           end
 
@@ -170,6 +208,14 @@ module OodCore
           # Password file for VNC server
           def vnc_passwd
             context.fetch(:vnc_passwd, "vnc.passwd").to_s
+          end
+
+          def container_module 
+            context.fetch(:container_module, "singularity").to_s
+          end
+
+          def container_command 
+            context.fetch(:container_command, "singularity").to_s
           end
 
           # Arguments sent to `vncserver` command
@@ -197,7 +243,7 @@ module OodCore
           # Clean up any stale VNC sessions
           def vnc_clean
             context.fetch(:vnc_clean) do
-              %(vncserver -list | awk '/^:/{system("kill -0 "$2" 2>/dev/null || vncserver -kill "$1)}')
+              %(#{container_command} exec instance://#{@instance_name} vncserver -list | awk '/^:/{system("kill -0 "$2" 2>/dev/null || #{container_command} exec instance://#{@instance_name} vncserver -kill "$1)}')
             end.to_s
           end
       end