+++ lib/ood_core/job/adapters/sge/qstat_xml_r_listener.rb
@@ -13,7 +13,6 @@ require 'date'
 # :queue_name
 # :status
 # :wallclock_limit
-
 class QstatXmlRListener
   # [Array<Hash>]
   attr_reader :parsed_jobs